get rid of float warning

+++ examples/volume_plane_rendering.py
@@ -3,6 +3,7 @@ Display one 3-D volume layer using the add_volume API and display it as a plane
 with a simple widget for modifying plane parameters
 """
 import napari
+import numpy as np
 from napari.utils.translations import trans
 from skimage import data
 
@@ -11,7 +12,7 @@ viewer = napari.Viewer(ndisplay=3)
 # add a volume
 blobs = data.binary_blobs(
     length=64, volume_fraction=0.1, n_dim=3
-).astype(float)
+).astype(np.float32)
 volume_layer = viewer.add_image(
     blobs, rendering='mip', name='volume', blending='additive', opacity=0.25
 )
